POST Record

+++ backend/src/main/java/com/devsuperior/dspesquisa/repositories/GameRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import com.devsuperior.dspesquisa.entities.Game;
 
 @Repository
-public interface GameRepository extends JpaRepository<Game, Integer> {
+public interface GameRepository extends JpaRepository<Game, Long> {
