+++ backend/src/main/java/com/devsuperior/dspesquisa/repositories/RecordRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.devsuperior.dspesquisa.entities.Record;
 
 @Repository
-public interface RecordRepository extends JpaRepository<Record, Integer>{
+public interface RecordRepository extends JpaRepository<Record, Long>{
 
 }
